+++ src/rfc9110/mod.rs
@@ -6,6 +6,7 @@
 //!
 //! [RFC 9110]: https://www.rfc-editor.org/rfc/rfc9110
 
+pub mod challenge;
 pub mod chars;
 pub mod headers;
 pub mod request;